core/pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "alltrue-customer-core"
dynamic = ["version"]
authors = [
    {name = "AllTrue.ai INC", email = "ury@alltrue.ai"}
]
description = "Client Library for AllTrue.ai API"
dependencies = [
    "hishel~=0.1.1",
    "httpx~=0.28.0",
    "logfunc~=2.9.1",
    "pydantic>=2.6.4",
    "python-dotenv~=0.21.1",
]
requires-python = ">= 3.11"

[project.urls]
Repository = "https://github.com/AllTrue-ai/ai-security-customer-plane"

[tool.pdm.version]
source = "file"
path = "../src/version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/alltrue"]
